+++ src/Products/PloneMeeting/ToolPloneMeeting.py
@@ -1141,10 +1141,12 @@ class ToolPloneMeeting(UniqueObject, OrderedBaseFolder, BrowserDefaultMixin):
                                   action_label=None,
                                   user_id=newOwnerId or newItem.Creator())
 
-            # The copy/paste has transferred annotations, we do not need them.
-            for ann in annotations:
-                if ann.startswith(SENT_TO_OTHER_MC_ANNOTATION_BASE_KEY):
-                    del annotations[ann]
+            # The copy/paste has transferred annotations,
+            # remove ones related to item sent to other MC
+            anns_to_remove = [ann for ann in annotations
+                              if ann.startswith(SENT_TO_OTHER_MC_ANNOTATION_BASE_KEY)]
+            for ann_to_remove in anns_to_remove:
+                del annotations[ann_to_remove]
 
             self.REQUEST.set('currentlyPastingItems', False)
         return newItem